feat(tracks): introduce comprehensive tracks management system
- Added new classes for tracks analysis, export, and indexing, enhancing the overall tracks management functionality.
- Implemented `TracksAnalyzer` for analyzing media tracks and retrieving stream data based on magnet links.
- Created `TracksExportService` to facilitate the export of track data, including validation and directory management.
- Developed `TracksIndexManager` for efficient indexing and persistence of track information.
- Introduced `TracksStatsCache` for caching statistics related to tracks, improving performance during data retrieval.
- Established `TracksLogging` for structured logging of track-related events and errors.
- Enhanced `TracksPathResolver` for improved path management and legacy file handling.

This update significantly improves the organization and functionality of the tracks module, paving the way for better data handling and user experience.
